sdk/autonomous-agent-starter/requirements.txt
# Core dependencies
requests>=2.28.0
pyyaml>=6.0

# LLM providers (install the one you need)
openai>=1.0.0
anthropic>=0.18.0
